requirements.txt
flask==3.0.3
ultralytics==8.1.0
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
timm==0.9.12
Pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0